requirements.txt
annotated-types==0.6.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
instagrapi==2.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
moviepy==1.0.3
mysql-connector-python==8.4.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proglog==0.1.10
psycopg2-binary==2.9.9
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.0
Werkzeug==3.0.3
